Add and improve endpoints

+++ src/main/java/com/kCalControl/controller/impl/UserDBControllerImpl.java
@@ -60,8 +60,8 @@ public class UserDBControllerImpl implements UserDBController {
     }
 
     @Override
-    public ResponseEntity<String> getLoggedUserData(ObjectId id, Principal principal) {
-        if(!checker.checkSameUser(id, principal)){
+    public ResponseEntity<String> getLoggedUserData(ObjectId id) {
+        if(!checker.checkSameUser(id)){
             return ResponseEntity.status(403).build();
         }
         UserDB userDB = userDBService.returnUserById(id);
@@ -71,14 +71,19 @@ public class UserDBControllerImpl implements UserDBController {
     }
 
     @Override
-    public ResponseEntity<Void> deleteUser(ObjectId id, Principal principal) {
-        checker.checkSameUser(id, principal);
+    public ResponseEntity<String> deleteUser(ObjectId id) {
+        if(!checker.checkSameUser(id)){
+            return ResponseEntity.status(403).build();
+        }
         userDBService.deleteUser(id);
-        return ResponseEntity.ok().build();
+        return ResponseEntity.ok("User deleted successfully");
     }
 
     @Override
-    public ResponseEntity<Void> updateUserData(ObjectId id, UpdateUserDataDTO dto) {
+    public ResponseEntity<String> updateUserData(ObjectId id, UpdateUserDataDTO dto) {
+        if(!checker.checkSameUser(id)){
+            return ResponseEntity.status(403).build();
+        }
         UserDB moddedUser = userDBService.returnUserById(id);
         UserDB modificationUser = userDBService.returnLoggedUser();
 
@@ -92,11 +97,14 @@ public class UserDBControllerImpl implements UserDBController {
         assetsRepository.save(moddedUser.getAssets());
         userDBRepository.save(moddedUser);
 
-        return ResponseEntity.ok().build();
+        return ResponseEntity.ok("User data updated successfully");
     }
 
     @Override
-    public ResponseEntity<Void> updatePassword(ObjectId id, UpdatePasswordDTO dto) {
+    public ResponseEntity<String> updatePassword(ObjectId id, UpdatePasswordDTO dto) {
+        if(!checker.checkSameUser(id)){
+            return ResponseEntity.status(403).build();
+        }
         UserDB moddedUser = userDBService.returnUserById(id);
         UserDB modificationUser = userDBService.returnLoggedUser();
 
@@ -107,6 +115,6 @@ public class UserDBControllerImpl implements UserDBController {
 
         assetsRepository.save(moddedUser.getAssets());
         userDBRepository.save(moddedUser);
-        return ResponseEntity.ok().build();
+        return ResponseEntity.ok("Password updated successfully");
     }
 }
